Spring security configured

+++ src/main/java/com/library/service/UserService.java
@@ -4,6 +4,10 @@ import java.util.HashSet;
 import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.library.exception.DataNotFoundException;
+import com.library.exception.UserRoleAssignException;
+import com.library.exception.UserRoleException;
 import com.library.model.Role;
 import com.library.model.User;
 import com.library.repository.RoleRepository;
@@ -21,20 +25,21 @@ public class UserService {
 	{
 		Set<Role> roles = new HashSet<>();
 		
-		roles.add(roleRepository.findById(2).get());
+		roles.add(roleRepository.findById(2).orElseThrow(() -> new DataNotFoundException()));
 		user.setRoles(roles);
 		userRepository.save(user);
 	}
 
 	public void deleteStudent(Integer studentId) {
-		
+		if(userRepository.findById(studentId).get() == null)
+		{
+			throw new DataNotFoundException();
+		}	
 		userRepository.deleteById(studentId);
-		
 	}
 	
 	public User updateStudent(User user)
 	{
-		
 		User userDb = userRepository.findByEmailid(user.getEmailid());
 		userDb.setFirstName(user.getFirstName());
 		userDb.setLastName(user.getLastName());
@@ -45,24 +50,26 @@ public class UserService {
 		
 	}
 	
-	public User removeRole(Integer id) throws Exception
+	public User removeRole(Integer id)
 	{
 		User user = userRepository.findById(id).get();
-		if(id == user.getId())
+		if(user.getRoles()==null)
 		{
+			throw new UserRoleException();
+		}
 		user.setRoles(null);
 		return userRepository.save(user);
-		}
-		else {
-			throw new Exception("User's role is removed already.") ;
-		}
 	}
 	
 	public User assignRole(Integer id)
 	{
 		User user = userRepository.findById(id).get();
-		Set<Role> roles = new HashSet<>();
+		if(user.getRoles()!=null)
+		{
+			throw new UserRoleAssignException();
+		}
 		
+		Set<Role> roles = new HashSet<>();
 		roles.add(roleRepository.findById(2).get());
 		user.setRoles(roles);
 		return userRepository.save(user);
